fix error: delete null avatar image

+++ user-service/src/main/java/com/onlinestories/user_service/Service/UserService.java
@@ -232,8 +232,10 @@ public class UserService {
             }
 
             if(file != null && !file.isEmpty()){
-                String message = mediaClient.deleteFile(user.getImg(), "image");
-                log.info("Deleted old avatar for userId {}: {}", userId, message);
+                if(user.getImg() != null) {
+                    String message = mediaClient.deleteFile(user.getImg(), "image");
+                    log.info("Deleted old avatar for userId {}: {}", userId, message);
+                }
                 String imgUrl = mediaClient.uploadFile(file, "avatars");
                 user.setImg(imgUrl);
             }
